Added files via upload
A Java program that allows user to write text into a textarea and edit their text by using the buttons in the toolbar, or by accessing Edit functions in the menubar. User can also load contents of an internal text file, and may save any text they type in the textarea into that text file as well.